Unify CI and CD pipelines
I split deployment into a pipeline separate from CI; the experience on GitHub isn't ideal, however, as the deployment is detached from the commit. As a result, the deployment status check doesn't appear alongside the commit: it wouldn't be so apparent if it were to fail.

This change moves deployment to the CI pipeline. It only occurs after the other jobs have succeeded and the commit is on the 'main' branch.

Refs f95109a700eaa4066a9d39594698ec9e42318750