Build an app
Instead of serving the static files through a CLI command, this change introduces a script built with TypeScript, which the container runs.

The container can run the 'node' command directly rather than 'npm'; if running the container locally, you might need to run it with the '--init' flag to handle SIGINT correctly.

Refs https://docs.docker.com/engine/reference/run/