Mock the connection to Zenodo
This change mocks the responses rather than connecting to the Zenodo sandbox in the integration tests.

As a result, we can now control what happens on the Zenodo end, rather than being subjected to unscheduled downtime and rate-limiting. It also gives us the chance to have predictable data.

This change is at the expense of knowing that it works when talking to an actual instance of Zenodo.

(I tried running Zenodo locally following their Docker guidance, but it doesn't work and doesn't seem to have for a long time.)

Refs #5, https://github.com/zenodo/zenodo/issues/1835