Turn the description into HTML
Zenodo expects the record description to be HTML, but unlike the editor on their upload page, the API won't transform plain text into HTML. So, this uses the Markdown-It library (https://markdown-it.github.io/) to parse the review as Markdown: this should be good enough for now.

As we use the description when displaying the review, this change will also help the preprint/review pages.

Refs #5