Expect request details beyond the URL in tests
We've recently changed some of the data that we send to Zenodo without having to change the tests. To avoid future problems, especially now the integration tests don't talk to a real Zenodo, this updates the mock request matches to include headers and bodies. When it doesn't match, the experience isn't great (ideally, it would produce a diff, but instead, you have to debug why it didn't find a match).

This commit is beginning to show that an abstraction layer would be helpful here (so this code doesn't have to know the details of the connection to Zenodo), but I'm resisting the urge for now.

Refs #30, 03f776fb9ecf9993afff2bfcb76182ce0c83a182, 2db0baf4e30e76a158fb33e7c94ef5e790ed65c3, c94f42676e4dc1c20a9a97c0677e09749097f9c2, https://www.wheresrhys.co.uk/fetch-mock/#usagedebug-mode