Simplify the lookup-doi route
Currently, the route appears both as having a query string (when accessing it) and not (in the form). This commit changes the form to a POST to keep it simpler. It also redirects invalid input (i.e. anything that isn't a DOI) to the homepage rather than showing the 'not found' message.