Run Playwright through Docker
Visual regression testing is hard cross-platform due to rendering differences, so the recommended approach is to run it inside Docker for consistency. So, as a prerequisite, this changes Playwright to run inside the official container.

In the future, it seems that it will be possible to run just the browsers inside Docker, which would be preferable.

Refs https://github.com/microsoft/playwright/issues/8161