Fix Docker caching
The 'docker build' command isn't using the Docker cache, so it's rebuilding the image (and is slow). I want to keep the separate CI step to be able to see the output; I think it's not using it due to the second 'docker build' not using Docker Buildx. So, this change will make 'docker build' an alias for 'docker buildx'.

Refs 5b596658874a07b82948e8e2fde1250faf703014, https://github.com/docker/setup-buildx-action#install-by-default