Add ORCID authentication
Uses the ORCID OAuth2 service to find the user's name and ORCID iD.

I'm not entirely happy with the 'redirectUri' setting: it feels like it should be possible to generate it from the request's URL, but I've not been able to see a way to do so. Setting it from the outside has caused a few changes (setting environment variables when running the app and knowing the port number in the integration tests before the server starts).

The integration tests use a publically-available MocK OAuth2 server and then separately mock the access token response to include the extra details that ORCID provides. Again, I'm not entirely happy with this, but it works.

Refs #29