Set up Webpack
This change uses Webpack (https://webpack.js.org/) to 'build' the assets. There's no benefit yet (as the asset files remain the same), but it allows us to manipulate them (e.g. )

I've chosen Webpack over other tools (Parcel, Vite etc.) as it looks to be the only one that can produce a sensible manifest file for the server (a map of input filenames to output filenames).