Remove knowledge of asset paths from the app
This change sees Webpack produce a JSON manifest file which the app can use to load the public path of asset files. As a result, we will be able to change the name of the generated files in the future (e.g. for cache-busting).

As TypeScript reads the JSON file, some CI tasks need to be aware of the manifest file.