Process imported CSS files using PostCSS
As it stands, PostCSS doesn't process CSS files imported by the stylesheet, which means we're unable to transform them. As a result, the modern CSS features used in the main stylesheet (e.g. nesting) are not available in the others.

This change configures Webpack's css-loader appropriately. I discovered, however, that it causes the postcss-custom-properties plugin to behave strangely. The plugin seems to be going through a significant change, and it's not clear/easy to understand how we should use it in this case. I've had to also stop PurgeCSS from removing unused font declarations, as it doesn't discover font names that now only appear in custom properties.

Refs https://webpack.js.org/loaders/css-loader/#importloaders, https://github.com/csstools/postcss-plugins/discussions/192