Use CSS Custom Selectors
This change adopts the CSS Custom Selectors draft specification (https://drafts.csswg.org/css-extensions/#custom-selectors) to reduce the amount of duplication with some selectors. For example, the various heading levels ('h1', 'h2' etc) are grouped in a ':--heading' selector.

I hope this is the right balance and isn't a hindrance when reading the source. I wondered whether to name the selectors like the ':any-link' pseudo-class (e.g. ':--any-heading'); this might be something to review.