Limit HTML type
When using Node.js (without a DOM shim), nanohtml returns a String rather than a DOM HTMLElement as defined in its types. According to its documentation, it could also return a DOM DocumentFragment, which again isn't defined in its types.

To avoid surprises, this change limits the Html type so that the internal details are hidden.