Make sure the app works with JavaScript disabled
While parts of the deployed app have a JavaScript dependency due to the ORCID log-in page (unfortunately), we have no guarantees that JavaScript is available on any of our pages. This change duplicates each Playwright environment with one that has JavaScript disabled to ensure that the app is still usable.

Refs #101, 2abd01f1bc144564034e3089f8de6bc3c0424505, https://kryogenix.org/code/browser/everyonehasjs.html