Add formatted text to be pasted in the PREreview input
This commit replaces the textarea with a Tiptap editor when JavaScript is available. I've tried to configure it the same as the sanitize-html library, but there's a potential for drift.

I chose Tiptap as it seems highly customisable and is based on ProseMirror (which has a good reputation).

There are some unfortunate differences in Playwright depending on whether it selects the textarea or the content-editable region (e.g. the former requires the toHaveValue assertion,  but the latter toHaveText). I also saw a race condition in Firefox between the JavaScript loading and Playwright trying to insert the text, so I had to add a step to wait for the content-editable region to be available if JavaScript is enabled.

Refs #101, https://tiptap.dev/, https://prosemirror.net/