Reduce specificity
PostCSS is transforming a :not() pseudo-class with multiple arguments into multiple :not() pseudo-classes. While this has wider browser support, it comes at the cost of higher specificity. As we have to set the outline-offset property on all :focus-visible() pseudo-classes to override browser defaults, this change removes the use of :not().

Refs #109