Remove redundant expectations
Given that we mock the Zenodo data in the integration tests, it's redundant to try and test it.

The partial screenshots are also quite volatile: other changes on the page cause whitespace differences on their edge, which isn't actually a change. I've not been able to work out why Playwright produces this.

Refs 802ebe1d7673d748d40c26717944d54af4f50d5d