Separate building the integration test image from running it
As the amount of integration tests increases, the time it takes increases. Running them inside a container in a GitHub action means they only have one worker available.

This change attempts to separate the Docker image building for the integration tests from running them. I'm trying this to see if it's practical to split the running of the tests over multiple jobs without having to build the image every time.

Given that the image contains both the app and the browsers, it's pretty large. The cost of exporting the image, uploading it, downloading it, and importing it might not be worth it.

Ideally, Playwright would only need to run the browsers in a container: it would simplify this step significantly, as CI would only have to download the image and build the app outside of a container.

Refs 861a110558ad8d624c847681a0b094a271aa9a26, https://github.com/microsoft/playwright/issues/8161, https://playwright.dev/docs/test-parallel#shard-tests-between-multiple-machines