Represent dates correctly
The Date API in JavaScript is pretty broken, and we're not referring to zoned date-times in preprints and reviews but plain dates.

This change introduces the Temporal proposal for its 'PlainDate' type and uses a polyfill. While this isn't ideal, the proposal seems stable for this usage and will provide a far better experience.

Refs #160, https://tc39.es/proposal-temporal/