Avoid some duplicate PREreviews
If the user manages to submit the post-your-PREreview form twice, the app relies on the form data having been deleted to avoid a duplicate. As posting the data to Zenodo can take a few seconds, this change deletes the form data beforehand (i.e. as early as possible) to avoid this scenario.

This change is imperfect as it's still vulnerable (only a lot less likely) and now loses the form data if posting the data to Zenodo fails.

Refs 4200628728d4b4f7d91956578a59b141c27a3ef4