Prevent the post-your-PREreview form from being submitted twice
As handling the form can take a few seconds due to interactions with Zenodo, this change prevents the form from being submitted a second time when JavaScript is available. It also disables the 'Post PREreview' button.

This change is far from ideal: a better solution is to decouple sending the data to Zenodo from handling the form submission and prevent duplicate PREreviews entirely. These changes are more involved, though.

Refs 4a9e896cfb2819f2f7fa22ee927319fb7fe504a7, https://design-system.service.gov.uk/components/button/#disabled-buttons,  https://axesslab.com/disabled-buttons-suck/