Avoid buttons being too long or overflowing
This change ensures the clickable area for a button isn't too small if the text is short.

It also removes the 'white-space' property, which caused it to overflow the page rather than wrap if it's more than a line long. I'm not sure why this property was there; I suspect I just copied it from elsewhere.

Refs https://rtlstyling.com/posts/rtl-styling#3.-differences-in-word-sizes-between-languages, https://ishadeed.com/article/css-short-long-content/