Retain original PREreview formatting
Zenodo heavily sanitises record descriptions (e.g. removing headings), which cuts out our formatting. Instead, we have to load the PREreview text from the HTML file and use that, as it contains the original HTML (warts and all).

This change means an extra HTTP request, but it should be possible to cache this heavily (as Zenodo don't allow changes to these files).

Refs #191