Require DOIs in URLs to be lowercase
DOIs are case insensitive, but URLs aren't. While it's possible to define canonical URLs that allow user agents to realise where there are duplicates, this change avoids the issue by requiring URL DOIs to be lowercase.

This change is a prerequisite for allowing DOIs that contain letters. (Neither bioRxiv nor medRxiv does so.)

Refs #249, https://doi.org/10.17487/RFC6596