Load data for SciELO preprints
This change allows the data for SciELO preprints to be loaded. The server is multilingual, and the Crossref API doesn't say which language the title and abstract are in, so it uses the TinyLD library to guess from the known possibilities.

While experimenting, I found that TinyLD could only work out the language correctly when restricting the possibilities. node-cld was able to determine the language accurately, but it's a large package, so I'm avoiding it now.

Refs #249, https://github.com/komodojp/tinyld, https://github.com/dachev/node-cld