Use replaceAll rather than regex to replace multiple values
'replaceAll' is available in Node.js but not in TypeScript, as setting the 'lib' option disabled the base types. This change restores the correct types and uses the 'replaceAll' function.

I only spotted this as my IDE (PhpStorm) was erroneously flagging 'Array.isArray' as deprecated, which appears to have another side effect of the typing problem.

Refs https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replaceAll, https://www.typescriptlang.org/tsconfig#extends, https://github.com/tsconfig/bases/blob/7be7a0937cb80761b9e1f64ec5c18c2aa0416530/bases/node16.json#L6, https://stackoverflow.com/questions/56210870/array-isarray-is-deprecated-now#comment128435830_56210870