Reuse known PREreview text
Zenodo doesn't set a 'Cache-Control' header when requesting files, which is a shame as they are immutable. This change uses the Fetch API's 'force-cache' mode so that if we already have a copy of the file, we don't fetch it again, reducing the number of HTTP requests we make.

The behaviour is similar to if Zenodo were to return an immutable HTTP response.

Refs #249, c192f8372b636a2b0c96a57efc0d700a01365b3d, https://help.zenodo.org/, https://doi.org/10.17487/RFC8246