Let the user know what to do if they don't have a pseudonym
If someone tries to log in to the new site without ever doing so on the old site, it will fail as they don't have a pseudonym generated. This change adds a message directing them to log in on the old site first.

I'm not convinced this message is perfect, but as the scenario is unlikely, it's probably not worth spending too much time on.

Refs #286, 2eff9a8d5a0f83deabd9574156d8d08edcbf2b24, 3b2a87e667af8cd08ca88ced38b0a831accf1958