Load data for AfricArXiv preprints
This change allows the data for AfricArXiv preprints to be loaded.

According to AfricArXiv, preprints can be in any language used in Africa, but there should be a summary in English, French, or both. I've been through the current preprints, and there aren't any other languages.

Some preprints, however, seem to have their title in French but the abstract in English. Others have abstracts in both languages but an unstructured format. I'm not sure if we can do anything about these.

Refs #339, https://info.africarxiv.org/languages/, https://doi.org/10.31730/osf.io/yv9az, https://osf.io/preprints/africarxiv/2csq6/