Retain invalid fields on the add-another-author form
If the ORCID iD is invalid, this change keeps it on the form so the user can edit it instead of reentering it.

I'm not too happy with the 'getInput' function. Still, having experimented with separating parsing from validating, it is easier to find the original value stored inside the decoder error.

Refs #319