Don't decode the add-an-author form twice
This change removes duplication so that instead of checking the form as a whole and checking each field individually, we derive the former from the latter. (This form isn't complicated in that one field doesn't depend on another, but there are ways to achieve that in this style.)

I'm not that happy with the way of working out if every field is valid (and then deriving the result if so): I tried to use 'sequenceS', but it has trouble merging different error types.

Refs #319, https://gcanti.github.io/fp-ts/modules/Apply.ts.html#sequences