Make the competing interests form errors typesafe
This commit makes the form similar to the add-an-author form. It's more complicated, however, as one field depends on the value of another. I've used pattern matching rather than a complicated decoder to determine whether to decode the conditional field.

I haven't captured the dependency in the CompetingInterestsForm type (e.g. if 'competingInterests' is 'no', 'competingInterestsDetails' can only be 'undefined'). It should be possible to do so, but I'm unsure if it would add anything.

Refs #319, a8746835d8422f08ec42c7fe950b72d46b7f978d