Add a link button
This change adds a button which allows setting a link using the browser's native prompt window.

While this works (and is accessible), it isn't the best experience, nor does it look good. It's a bit of a rabbit hole to implement a UI for inputting a URL; other editors (TinyMCE, CKEditor, Trix etc.) all take different approaches. So, for now, this is enough.

Also, this doesn't yet allow for editing the link URL: you have to remove it and add a new one.

Refs #431, https://developer.mozilla.org/en-US/docs/Web/API/Window/prompt