Ask if you have already written your PREreview
We believe that most people who post a PREreview have already written it in Google Docs or similar. They are pasting rather than writing directly in our form.

This change adds a step before the 'Write your PREreview' form that asks if you have already written it. It doesn't change anything yet but is a prerequisite to customising the flow (e.g. changing the wording from 'Write' to 'Paste').

I've not included it as a regular step (i.e. it's not one that you can return to), so it doesn't have its own URL.

Refs #432