Remove a possible race condition
I've noticed that the integration test periodically fails on CI with Safari, as the page scrolls down slightly when we take a snapshot. I've struggled to recreate this locally, only having one failure in 100 runs.

This commit removes a manual focus, hopefully making it more predictable. (Clicking the 'Heading level 1' button gives focus to the text editor, so it's redundant to do so beforehand.)

Refs #432