Present the PREreview text as Markdown when JavaScript is not available
Rather than presenting HTML to a user who doesn't have the HTML editor available, this change converts it to Markdown and shows that instead. We keep the HTML source in a separate (hidden) textarea, and the JavaScript loads that instead. I feel like the implementation is a bit of a hack, but it avoids having to turn the Markdown back into HTML when loading the editor.

Refs #441