Guess the language of SciELO preprint titles
I spotted a SciELO preprint in the Crossref API with its title in a different language than the abstract. So, this change guesses the language for SciELO preprint titles (like it does for AfricArXiv preprints).

This change also removes the fallback in favour of separating the title and abstract languages.

Refs https://api.crossref.org/works/10.1590/scielopreprints.5051