Try and make a test deterministic
We've continued to see failures in CI where the page scrolling is different in Safari; I can still replicate this locally, though at a lesser frequency. I'm still unsure what's happening; I wonder if it's to do with Safari's lack of support for the 'overflow-anchor' property, but that doesn't explain why it produces differing results. This change adds an explicit scroll in Safari, which has led to it passing 150 times locally.

Refs 68053233db02f95fa2bb66cffcbef74df25f895d, https://github.com/ProseMirror/prosemirror/issues/1169, https://developer.mozilla.org/en-US/docs/Web/CSS/overflow-anchor