Remove a race condition
I saw a test fail as the browser located the textarea before the HTML editor loaded but timed out waiting as the editor immediately hid it. This change makes it wait for the editor to load before locating the right textarea to edit.

This behaviour raises a few interesting questions. It feels like Playwright should keep trying to find an element that matches rather than doing it once and then waiting for it to be visible. Also, it raises the question of what should happen if the user edits the textarea before the HTML editor is loaded: currently, the HTML editor will replace it.