Avoid hard-coding the ORCID callback path
The mock OAuth server we're currently using doesn't provide a way to trigger an error, so we have to visit the path manually. We can reduce this impact by using the redirect URI stored on the page, so the only part the integration test has to know is the OAuth error redirect format. This still isn't ideal, but without it having a 'Deny Access' button, it's the best we can do.

Refs #566