Store sessions in Redis
To deploy more than one instance of the app, we need to be able to replicate the sessions. Otherwise, you could appear as logged out if a different one serves your request. I'm interested in using LiteFS to replicate the SQLite database, but it's not yet stable enough to try out. Fly.io have a Redis service that we can use instead, and as the current state is all key-value pairs, this fits well.

I've made the REDIS_URI environment variable optional, so we can run it locally without running a Redis instance. If it's not present, we use the in-memory Keyv adapter instead. (So the data will not survive the process restarting.)

This change does mean that we log everyone out, but there will be no data loss (unless someone happens to be submitting a form at the time).

Refs #568, https://fly.io/docs/reference/redis/, https://fly.io/docs/litefs/