Stop the session data from constantly growing
This change adds a time-to-live value (TTL) to each session entry so that they are removed after 30 days. The session cookie only lasts for the browser session, so this change will only log someone out if they leave their browser open for the duration.

Adding a TTL wasn't possible using the Keyv SQLite adapter.

Also, using a TTL means that if keys have to be evicted, the Fly.io Redis instance will evict sessions rather than the more-permanent form data. (Note that this option is currently disabled, as it's unlikely to be required for a long, long time.)

Refs #568, f64102eb65c7d0951a0c85d5d20cbc451b032cdc, https://fly.io/docs/reference/redis/#memory-limits-and-object-eviction-policies