Acknowledge that other authors may not want to be listed
The question 'Did you review this preprint with anyone else?' doesn't necessarily result in more authors, as they may not want to be listed. This change adds another option that covers this (and rewords the others to match).

We've been through a few variations on the wording, and this seems like the best balance. But this is something we'll need real-world feedback on.

Refs #587, https://www.gov.uk/service-manual/design/designing-good-questions#change-the-questions-until-you-find-what-works