Ensure the user is logged out after posting a PREreview
Browsers use a back/forward cache to help improve page loading times when using the back and forwards buttons. When we log the user out, we need to make sure that pressing the back button doesn't then show the user as being logged in again. We can do this by setting a 'Vary: Cookie' header so that the browser won't reuse the previous page as we've removed the cookie. Usually, this header is bad for HTTP caching rates (when combined with other services that set cookies, such as analytics services), but as we're not caching in a proxy, it's OK to use.

Refs #589, https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Vary, https://web.dev/bfcache/