Prevent 'change' links from appearing as disabled
The grey secondary button style has been confused with a button that has been disabled. This change removes the button styling entirely for the 'Change PREreview' and 'Change name' buttons, as they are links rather than buttons. While they could probably do with further improvement, hopefully, this is enough to avoid confusion.

Since the links are not part of a paragraph of text, I've left a bit of padding to make the target area slightly more prominent.

Refs #603, https://prereview.slack.com/archives/CR6B5804W/p1673486274003519?thread_ts=1673461805.675519&cid=CR6B5804W