Stop Safari from wrapping the text on buttons
It seemed not to happen in Playwright, but Safari on both my Mac and iPhone wrapped text on smaller buttons when they had wrapping whitespace.

The other options I looked at were not allowing the text to wrap in buttons (which isn't appropriate for long buttons or large text sizes), or using an inline flexbox (which can cause other issues, especially when mixing text nodes and elements).

Refs #603, https://prereview.slack.com/archives/CR6B5804W/p1673486274003519?thread_ts=1673461805.675519&cid=CR6B5804W