Get around IPv6 problems when using the Zenodo sandbox
We have strange connection problems when using the Zenodo sandbox. The only difference I can see using cURL is that it tries IPv6 and then falls back to using IPv4 (the production Zenodo uses IPv4 straightaway). I'm no expert on this, but I believe this means there's a configuration problem on Zenodo's end. The next version of Node has a new option that implements the Happy Eyeballs algorithm, which I believe would fix this. In the meantime, this change forces the use of IPv4 when using the Zenodo sandbox.

Refs https://en.wikipedia.org/wiki/Happy_Eyeballs, https://github.com/nodejs/node/issues/40702#issuecomment-958380640, https://nodejs.org/api/net.html#netsetdefaultautoselectfamilyvalue