Reduce the number of HTTP requests to delay hitting API rate limits
Suppose two visitors load the same preprint page at the same time. In that case, the app will request all the information (preprint details, Rapid PREreviews and Full PREreviews) twice. Caching helps, but we should still try and remove duplicates as we also make revalidation requests. This change introduces basic request collapsing: if we make two GET requests to the same URL, we only send one, and the response is shared.

I've thought that maybe we should switch to using a proxy that can handle these details for us, but that's a more significant change that requires investigating.

Refs #634, https://developer.mozilla.org/en-US/docs/Web/HTTP/Caching#request_collapse