Intentionally don't transform CSS custom properties
I noticed that PostCSS wasn't transforming most CSS custom properties because it didn't follow the imports to resolve them. However, it changed those declared in the main stylesheet (style.css); previously, it was disabled.

This change explicitly disables the plugin to match the previous behaviour and uses the documented workaround for following the imports.

Refs #616, https://github.com/csstools/postcss-plugins/blob/5c7e4d71ffc2264fa4dcb46d041b21312c9ea5d0/plugins/postcss-custom-properties/README.md#modular-css-processing