Try and unbreak Dependabot
I've noticed that Dependabot has not opened any pull requests recently; looking at its logs, I see npm failing due to the engine version check. The failure seems to have started with the move to npm 9. This change relaxes the constraint again to support both npm 8 and 9, which is hopefully enough for Dependabot to work.

Refs #635, https://github.com/dependabot/dependabot-core/issues/4072