Skip flaky high-contrast test assertions
When we select a text input and take a screenshot with high-contrast mode enabled, it's impossible to prevent the caret from appearing. This led to inconsistent snapshots and many test failures.

I looked at using Firefox instead, as it never includes the caret in screenshots. However, it doesn't support the mobile size, and the forced-color feature did not work correctly on the desktop.

I thought about bypassing the snapshots by deselecting the input (e.g. by pressing the 'Tab' key), but this conditionally changed the behaviour of the test and felt fragile.

So, I've skipped the test at the points where it's flaky. This is a little extreme, but it is honest about what's happened.

Refs #733, https://github.com/microsoft/playwright/issues/15211