Cater for slight rendering differences
I've noticed that the Chrome integration tests seem to have slight differences in the size of buttons, precisely the amount of block-end padding. I'm guessing this is due to inconsistent rounding; this is surprising, but rather than change the design, this commit allows for the snapshot comparisons to have a certain amount of difference. I've tried to be quite conservative in this to avoid mistakes.

Refs #801, https://playwright.dev/docs/test-snapshots