Prevent Chromium crashing
We've seen quite a few timeouts in Chromium integration tests recently. After some debugging, I discovered the browser itself was crashing. The stack trace shows that it is related to CSS animations. There seems to be a bug in how Playwright disables CSS animations when taking screenshots.

This commit uses the prefers-reduced-motion CSS media feature to disable animations entirely, mitigating the issue. (This is only possible as our animation only appears when it is set to 'no-preference'.)

Refs #801, https://github.com/microsoft/playwright/issues/15560, https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-reduced-motion