Run Prettier in parallel
Locally we have extra directories created by our IDEs, but we have Git configured globally to ignore these. Prettier will run on these directories and try to format the XML files. Prettier doesn't have the same concept of a global ignore file; we looked at passing it both the local and global .gitignore files, but it doesn't yet support this (it will in the next major version). In the short-term, we've had to add the directories created by our IDEs to Prettier's ignore file.

Refs #905, https://github.com/prettier/prettier/issues/8048