Remove the number of distractions on the write-a-PREreview flow
Adding links to the navigation bar causes us to regenerate all the screenshots; this takes quite a while. While we can block out parts of the screen to reduce the number of changes or reduce the number of screenshots entirely, I've wondered we need if things like the menu on these pages. As they're part of a workflow, we don't expect a user to go and look at the training that PREreview offers. So, this commit streamlines these pages by removing superfluous details like menus.

Refs #930, 8f216c40d023e448944c1d688f13910efd0440ab