Set a minimum size for the logos
Firefox wouldn't display the logos, so we had to set a minimum size in CSS. This change adds explicit dimensions in the SVG itself, much larger than we need, so they will always be scaled down to the size we want. It's still a hack, but less of a hack than we had previously.

Refs #931