Ensure that mocked requests are matched
This change uses fetch-mock's 'done' inspection to ensure that it matches the requests as expected to avoid tests passing for the wrong reason (i.e. we didn't mock a response).

I think I've found a bug in fetch-mock where it doesn't recognise that it matched a request when testing a query string independently of the URL. It works as expected when only partially matching the URL.

Refs #976, 047d979af39d6704b2ae42e1d1ff9caa8f277800, https://www.wheresrhys.co.uk/fetch-mock/#api-inspectiondone, https://github.com/wheresrhys/fetch-mock/issues/657