Expose heading-level issues in external content
Things like the review text and preprint abstract have headings that don't match the heading level where we display them. For example, it might contain an h1 element, and we include it on a page with an h1: the result confuses assistive technologies and crawlers and looks ugly. This commit adds headings to mock content to show the problem.

Refs #1027