Update heading levels to fit our use
This commit rewrites the heading levels in external content to match where we're using it, to prevent confusing assistive technologies and crawlers, and to look nicer. It uses regex to find the highest level in a piece of content and updates them based on the current level. Using a regex on HTML tends to be a bad idea, but the usage here is so limited that I don't believe it should be a problem. (Famous last words...)

Refs #1027, 6765dece8d13c093acf1de41d26811ec9b4716b7