Make it more apparent that there are more reviews than just those shown
I looked at including it as a final list item, but it didn't look great when wrapped at larger sizes. I also thought about only showing the link if there are more than 5 reviews: given this is the case in production, I took the more straightforward option of always displaying the link. This decision means it links to the first page rather than the second, which is something we could improve.

Refs #1089