Pin the module resolution strategy
The @tsconfig/node18 library is moving to use the 'node16' option, which is supposedly correct for "modern" code. However, like every other time I've tried to do anything related to ES Modules, things break, and I'm yet to find the bottom of the rabbit hole. This commit pins the option to the one that works.

Refs #1054, https://www.typescriptlang.org/tsconfig#moduleResolution