Prevent KaTeX from logging to the console
I've noticed that some generated text causes KaTeX to log to the console as it's not valid LaTeX (primarily when it contains a `%` character, which triggers a comment, but there's no newline at the end. By default, KaTeX handles these cases but warns about them. This commit turns off those warnings.

I looked at using the 'jest-fail-on-console' library to ensure that dependencies don't output things unexpectedly, but it's currently incompatible with our Jest configuration.

Refs #1124, https://katex.org/docs/options.html, https://github.com/ValentinH/jest-fail-on-console/issues/41