Add comment fields for each data-presentation response
This change adds a place to store a comment for each possible response (except for skipping it) to the data-presentation question. I've played around with various options, and although this is pretty verbose, it matches what I'm trying to achieve: a way to store a comment on each that we don't throw away if you change your response. This behaviour matches the rest of the form: if you've filled something out that no longer applies, you should be able to return to it by changing other answers. It doesn't apply to the CompletedForm type, however. So I've changed this type only to have the applicable comment.

Note that this commit doesn't provide a UI for writing a comment.

Refs #1135