Stop the TypeScript compiler from being overloaded
When I tried adding another set of comment boxes to a question field, the TypeScript compiler suddenly took a couple of minutes longer to compile the project. This change also affected ESLint and my IDE. This commit removes the use of 'sum' types for converting the question response and comment in favour of a manual approach, which seems to reduce the load.

Refs #1135