Find the image for an ORCID iD using the Cloudinary API
This change uses structured metadata to associate each image with an ORCID iD and then searches for the most recent one. This technique allows us to remove the hard coding without storing the association locally.

I've chosen to remove the images from the integration tests to avoid potential breaks in the future. We will revisit this when we have a way to upload an image.

Refs #1182, https://cloudinary.com/documentation/dam_admin_structured_metadata, 0f6aed1b2e474e7a68881a6da8664bf182caa65b